NOISSUE: add const for initialBalance in tests

+++ ledger-core/v2/application/functest/api_test_utils.go
@@ -150,8 +150,8 @@ func getWalletBalance(url, ref string) (int, error) {
 }
 
 // Adds amount to wallet.
-func addAmountToWallet(url, ref string, amount uint) error {
-	rawResp, err := sendAPIRequest(url, walletAddAmountRequestBody{To: ref, Amount: amount})
+func addAmountToWallet(url, ref string, amount int) error {
+	rawResp, err := sendAPIRequest(url, walletAddAmountRequestBody{To: ref, Amount: uint(amount)})
 	if err != nil {
 		return errors.Wrap(err, "failed to send request or get response body")
 	}
